backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
openai==1.51.0
fal-client==0.5.0
httpx==0.27.2
python-dotenv==1.0.1
Pillow==10.4.0
pydantic==2.9.2
aiofiles==24.1.0
ffmpeg-python==0.2.0
